refactor(experimental): reorganize swarm team into an experimental module
